refactor: relocate `EditTournament` and `DeleteTournament` commands/handlers/validators to `Domain` layer
- Moved `EditTournamentCommand`, `EditTournamentCommandHandler`, `EditTournamentCommandValidator`, `DeleteTournamentCommand`, `DeleteUserCommandValidator`, and their handlers to the `Domain` layer.
- Updated import paths across affected files for consistency and alignment with project structure.
