+++ tournament/src/Domain/Command/Validators/DeleteUserCommandValidator.ts
@@ -1,7 +1,7 @@
-import {BaseValidator} from "./BaseValidator.js";
-import {TournamentRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/TournamentRepository.js";
-import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
-import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog.js";
+import {BaseValidator} from "./BaseValidator";
+import {TournamentRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/TournamentRepository";
+import {NotificationError} from "../../../Shared/Errors/NotificationError";
+import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog";
 import {DeleteTournamentCommand} from "../CommandObject/DeleteTournamentCommand";
 
 export class DeleteUserCommandValidator implements BaseValidator<DeleteTournamentCommand>
