+++ tournament/src/Domain/Command/Validators/EditTournamentCommandValidator.ts
@@ -1,9 +1,9 @@
-import {BaseValidator} from "./BaseValidator.js";
-import {TournamentRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/TournamentRepository.js";
-import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
-import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog.js";
-import {CustomError} from "../../../Shared/Errors/CustomError.js";
-import {ValidationException} from "../../../Shared/Errors/ValidationException.js";
+import {BaseValidator} from "./BaseValidator";
+import {TournamentRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/TournamentRepository";
+import {NotificationError} from "../../../Shared/Errors/NotificationError";
+import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog";
+import {CustomError} from "../../../Shared/Errors/CustomError";
+import {ValidationException} from "../../../Shared/Errors/ValidationException";
 import {EditTournamentCommand} from "../CommandObject/EditTournamentCommand";
 import {BackendApiClient} from "../../../Infrastructure/Http/Concrete/BackendApiClient";
